requirements.txt
tablib[all]==3.5.0
PyYAML==6.0